backend/requirements.txt
Flask==3.1.1
Flask-CORS==5.0.1
Flask-SQLAlchemy==3.1.1
google-genai==1.14.0
python-dotenv==1.1.0
PyJWT==2.10.1
Werkzeug==3.1.3
